[SEPAB-8, SEPAB-21, SEPAB-23] Rework of all Routes and Controllers
