+++ backend/server/controllers/doctorController.js
@@ -1,0 +1,119 @@
+const asyncHandler = require('express-async-handler')
+const doctorModel = require('../models/doctor')
+
+// GET - Doctor
+const getDoctor = asyncHandler(async(req, res) => {
+
+    const doctorFirstName = req.body.drFirstName
+    const doctorLastName = req.body.drSurName
+
+    const doctor = await doctorModel.find({
+        drFirstName: doctorFirstName,
+        drSurName: doctorLastName
+    })
+
+    if (!doctorFirstName) {
+        res.status(400)
+        throw new Error(`Invalid Doctor Details: Missing inputs found in the request!`)
+    }
+
+    res.status(200).json(doctor)
+})
+
+// POST - Doctor
+const addDoctor = asyncHandler(async(req, res) => {
+
+    const doctor = await doctorModel.create({
+        drFirstName: req.body.drFirstName,
+        drSurName: req.body.drSurName,
+        drCode: req.body.drCode,
+        prescriberCode: req.body.prescriberCode,
+        drAddress: req.body.drAddress,
+        drPhoneNo: req.body.drPhoneNo,
+        drEmail: req.body.drEmail,
+        drClinic: req.body.drClinic
+    })
+
+    if (!doctor) {
+        res.status(400)
+        throw new Error(`Invalid Doctor Details: Missing inputs found in the request!`)
+    }
+
+    res.status(200).json(doctor)
+})
+
+// PUT - Doctor using ID
+const modifyDoctorByID = asyncHandler(async(req, res) => {
+    
+    const updatedDoctor = await doctorModel.findByIdAndUpdate(
+        req.params.id,
+        req.body, {
+            new: true,
+        }
+    )
+
+    if (!updatedDoctor) {
+        res.status(400)
+        throw new Error(`Invalid Doctor Search for update!`)
+    }
+
+    res.status(200).json(updatedDoctor)
+})
+
+// PUT - Doctor using FirstName LastName
+const modifyDoctorByName = asyncHandler(async(req, res) => {
+
+    const drFirstName = req.body.drFirstName
+    const drLastName = req.body.drSurName
+    
+    const updatedDoctor = await doctorModel.findOneAndUpdate(
+        { 
+            drFirstName: drFirstName, 
+            drLastName: drLastName 
+        },
+        req.body, {
+            new: true,
+        }
+    )
+
+    if (!updatedDoctor) {
+        res.status(400)
+        throw new Error(`Invalid Doctor Search for update!`)
+    }
+
+    res.status(200).json(updatedDoctor)
+})
+
+// DELETE - Doctor using ID
+const deleteDoctorByID = asyncHandler(async(req, res) => {
+    const doctor = await doctorModel.findById(req.params.id)
+
+    if (!doctor) {
+        res.status(400)
+        throw new Error(`Invalid Doctor Search for update!`)
+    }
+
+    await doctor.remove()
+    res.status(200).json(doctor)
+})
+
+// DELETE - Doctor using Name
+const deleteDoctorByName = asyncHandler(async(req, res) => {
+    const drFirstName = req.body.drFirstName
+    const drLastName = req.body.drSurName
+    const doctor = await doctorModel.findOneAndRemove({
+        drFirstName: drFirstName,
+        drLastName: drLastName
+    })
+
+    if (!doctor) {
+        res.status(400)
+        throw new Error(`Invalid Doctor Search for delete!`)
+    }
+
+    res.status(200).json(doctor)
+})
+
+module.exports = {
+    getDoctor, addDoctor, modifyDoctorByID, modifyDoctorByName, deleteDoctorByID, deleteDoctorByName
+}
